Create Red and White Wine ML Analysis
“Wine quality” dataset from the UCI Machine Learning Repository was chosen for our analysis and this dataset is related to red and white wine variants of Vinho Verde wine originating from Portugal in 2009. This dataset contains 4898 observations in total with 11 input variables and 1 output variable.

Research objective: to classify the quality of white wine against provided variables so that a common metric can be employed and provided to consumers for a simplified choosing process. 

I have targeted 2 main hypotheses with our research objective, which is that firstly, the quality of wine can be simplified to ‘good’ and ‘poor’ where kNN classification and logistic regression can be used to classify the white wine into these simplified groups. Secondly, since the flavour, scent and alcohol level of wines are determined by its concentration of chemical elements, there could be an alternative way for consumers to select wine other than through “good” or “poor” absolutes which is to be found through clustering.